+++ recommendation-and-troubleshooting/recommendation-and-troubleshooting/src/main/java/project/recommendationandtroubleshooting/repository/BugHistoryRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import project.recommendationandtroubleshooting.model.troubleshooting.BugHistory;
 
 @Repository
-public interface BugHistoryRepository extends JpaRepository<BugHistory, Long> {
+public interface BugHistoryRepository extends JpaRepository<BugHistory, Integer> {
 
 }
