+++ recommendation-and-troubleshooting/recommendation-and-troubleshooting/src/main/java/project/recommendationandtroubleshooting/repository/ConfigurationCharacteristicTypeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import project.recommendationandtroubleshooting.model.recommendation.ConfigurationCharacteristicType;
 
 @Repository
-public interface ConfigurationCharacteristicTypeRepository extends JpaRepository<ConfigurationCharacteristicType, Long> {
+public interface ConfigurationCharacteristicTypeRepository extends JpaRepository<ConfigurationCharacteristicType, Integer> {
 
 }
